Allow decimal precision stat values

+++ database/migrations/2018_12_20_022726_create_user_stats_table.php
@@ -17,7 +17,7 @@ class CreateUserStatsTable extends Migration
             $table->increments('id');
             $table->unsignedInteger('user_id');
             $table->string('key');
-            $table->bigInteger('value')->default(0);
+            $table->float('value')->default(0);
             $table->timestamps();
         });
     }
